+++ lib/src/reqif/reqif_spec_type.dart
@@ -40,6 +40,8 @@ class ReqIfSpecType extends ReqIfIdentifiable {
           children.add(ReqIfAttributeDefinition.parse(element, document, i));
         case "ATTRIBUTE-DEFINITION-STRING":
           children.add(ReqIfAttributeDefinition.parse(element, document, i));
+        case "ATTRIBUTE-DEFINITION-INTEGER":
+          children.add(ReqIfAttributeDefinition.parse(element, document, i));
         // TODO
 //        case "ATTRIBUTE-DEFINITION-BOOLEAN": children.add(ReqIfAttributeDefinition.parse(element,ReqIfElementTypes.datatypeDefinitionBoolean,"DATATYPE-DEFINITION-BOOLEAN-REF",document));
 //        case "ATTRIBUTE-DEFINITION-DATE": children.add(ReqIfAttributeDefinition.parse(element,ReqIfElementTypes.datatypeDefinitionDate,"DATATYPE-DEFINITION-DATE-REF",document));